requirements.txt
accimage==0.2.0
lmdb==1.3.0
matplotlib==3.5.3
mmcv_full==1.7.1
numpy==1.21.6
numpy==1.21.5
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
Pillow==9.2.0
Pillow==10.1.0
PyYAML==6.0
PyYAML==6.0.1
scipy==1.7.3
seaborn==0.13.0
tensorboardX==2.2
tensorboardX==2.6.2.2
torch==1.12.1
torchvision==0.13.1
